CR suggestions
