+++ .changeset/gold-masks-explain.md
@@ -1,0 +1,5 @@
+---
+"saleor-dashboard": minor
+---
+
+Refund add dialog no longer allows proceeding to refund with lines view when all lines have been refunded and manual refund view when user doesn't have payment permissions.
