--- .changeset/soft-wombats-fold.md
@@ -1,5 +1,0 @@
----
-"saleor-dashboard": minor
----
-
-This change disallows creating standard refund when all items have been already refunded and disallows creating manual refund when user has no permissions to do so.
